chore: relocate controller about auth, user
