+++ moit-api/src/main/kotlin/com/mashup/moit/controller/sample/SampleNotificationController.kt
@@ -1,4 +1,4 @@
-package com.mashup.moit.controller
+package com.mashup.moit.controller.sample
 
 import com.mashup.moit.common.MoitApiResponse
 import com.mashup.moit.infra.fcm.FCMNotificationService
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.RestController
 
 @RestController
 @RequestMapping("/api/v1/sample-noti")
-class NotificationController(
+class SampleNotificationController(
     private val fcmNotificationService: FCMNotificationService
 ) {
     @PostMapping("/pushs/topics")
